+++ tools/ghostftp-installer/main.go
@@ -18,7 +18,7 @@ import (
 	"time"
 )
 
-const version = "2.1.1-rc.19"
+const version = "2.1.1-rc.20"
 
 //go:embed site/* site/assets/* payload/GhostFTP.exe
 var bundle embed.FS